move BaseSchedule to same package as its subclasses
