+++ test/com/arpnetworking/metrics/portal/scheduling/impl/BaseScheduleTest.java
@@ -1,6 +1,5 @@
 package com.arpnetworking.metrics.portal.scheduling.impl;
 
-import models.internal.scheduling.BaseSchedule;
 import org.junit.Test;
 
 import java.time.Duration;
